refactor: organize code into modular structure
- Split monolithic code into separate files and modules
- Create dedicated services for tag generation, notifications, and note handling
- Implement proper typings and interfaces
- Add proper error handling and validation utilities
- Improve notification system with progress tracking
- Add proper project structure with src/ directory
- Add lint scripts to package.json
